# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /developer/releases:
    get:
      operationId: listDeveloperReleases
      tags: [developer]
      summary: List developer releases
      parameters:
        - $ref: "#/components/parameters/CollectionLimit"
        - $ref: "#/components/parameters/CollectionOffset"
      responses:
        "200":
          description: Page of releases, as an array
          content:
            application/json:
              schema:
                type: array
                items: { type: object }
...
components:
  parameters:
    CollectionLimit:
      name: limit
      in: query
      description: Number of records to return.
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    CollectionOffset:
      name: offset
      in: query
      description: Number of records to skip.
      schema: { type: integer, minimum: 0, maximum: 2147483647, default: 0 }
...